Extensive commit including a lot of sbt and automatical things we can use with jenkins to do integration tests. 1. Using process-runner https://github.com/whysoserious/sbt-process-runner to handle restapi activation project api > process-runner:status restapi [info] restapi is running. [success] Total time: 0 s, completed 5-gen-2015 0.35.29 > process-runner:stop restapi [info] restapi was stopped. Exit code: 1. [success] Total time: 0 s, completed 5-gen-2015 0.35.38 > process-runner:status restapi [info] restapi is idle. [success] Total time: 0 s, completed 5-gen-2015 0.35.42 >process-runner:start restapi ((( now api are UP and running ))) Relevant sources for this are api/plugins.sbt, api/project/Build.scala
2. tweaked a bit assembly to be able to produce Karedo.jar under api/target/Scala-2.11/
Main class is set to Rest so that you can easy launch this using java -Dconfig.resource=dummy.deployment.conf/-Dconfig.file=whatever Karedo.jar
Dummy is now automatically referring to port 8090 (to avoid conflicts with Jenkins) and to port 12345 for MongoDb testing as usual

3. Python has been tweaked to return exit code <> 0 if tests are failing, and to proper access port 8090 for testing.
NB: to have these tests working you must have python3 in the PATH (you can make an alias since many Unix distributions are considering python to be the old 2.7 versin and not the new one 3.x which is much better for our purposes).

So now the integration test can be achieved in the following way:
#sbt
>project api
>assembly // this will produce the Karedo.jar
>test // to do UnitTests with Mongodb possibly embedded on port 12345 or liveone if already up
>process-runner:start restapi
>python

What is left to do for complete integration with jenkins is
1. use nosetest PIP package for python3 which is able to produce junit xml that jenkins can understand
2. setup the Jenkins machine to have the python3 and related packages to be used with python tests
2. complete the Jenkins job so that will call sbt with proper sequence of commands as declared before..

NB: I tested so far with Windows, will test again on production test machine and on my virtual Ubuntu machine. It might not work with Mac...